rename 'cluster' in filenames
