+++ src/components/RuntimeManager.js
@@ -4,25 +4,22 @@ import _ from 'lodash/fp'
 import PropTypes from 'prop-types'
 import { Fragment, PureComponent, useState } from 'react'
 import { div, h, img, p, span } from 'react-hyperscript-helpers'
-import { GalaxyLaunchButton, GalaxyWarning } from 'src/components/cluster-common'
 import { ButtonPrimary, Clickable, IdContainer, Link, spinnerOverlay } from 'src/components/common'
 import { icon } from 'src/components/icons'
 import Modal from 'src/components/Modal'
-import { NewRuntimeModal } from 'src/components/NewClusterModal'
 import { NewGalaxyModal } from 'src/components/NewGalaxyModal'
+import { NewRuntimeModal } from 'src/components/NewRuntimeModal'
+import { GalaxyLaunchButton, GalaxyWarning } from 'src/components/runtime-common'
 import { dataSyncingDocUrl } from 'src/data/machines'
 import galaxyLogo from 'src/images/galaxy.svg'
 import rLogo from 'src/images/r-logo.svg'
 import { Ajax } from 'src/libs/ajax'
 import { getDynamic, setDynamic } from 'src/libs/browser-storage'
-import {
-  appIsSettingUp, collapsedRuntimeStatus, currentApp, currentRuntime,
-  persistentDiskCost, runtimeCost, trimRuntimesOldestFirst
-} from 'src/libs/cluster-utils'
 import colors from 'src/libs/colors'
 import { reportError, withErrorReporting } from 'src/libs/error'
 import * as Nav from 'src/libs/nav'
 import { clearNotification, notify } from 'src/libs/notifications'
+import { appIsSettingUp, collapsedRuntimeStatus, currentApp, currentRuntime, persistentDiskCost, runtimeCost, trimRuntimesOldestFirst } from 'src/libs/runtime-utils'
 import { errorNotifiedRuntimes } from 'src/libs/state'
 import * as Utils from 'src/libs/utils'
