+++ src/libs/routes.js
@@ -2,7 +2,7 @@ import _ from 'lodash/fp'
 import { compile, pathToRegexp } from 'path-to-regexp'
 import { routeHandlersStore } from 'src/libs/state'
 import * as Projects from 'src/pages/billing/List'
-import * as Environments from 'src/pages/Clusters'
+import * as Environments from 'src/pages/Environments'
 import * as Group from 'src/pages/groups/Group'
 import * as Groups from 'src/pages/groups/List'
 import * as HoF from 'src/pages/HoF'
