+++ src/pages/Main.js
@@ -4,7 +4,6 @@ import { hot } from 'react-hot-loader/root'
 import { h } from 'react-hyperscript-helpers'
 import ReactNotification from 'react-notifications-component'
 import AuthContainer from 'src/components/AuthContainer'
-import { AuthenticatedCookieSetter } from 'src/components/cluster-common'
 import ConfigOverridesWarning from 'src/components/ConfigOverridesWarning'
 import ErrorWrapper from 'src/components/ErrorWrapper'
 import Favicon from 'src/components/Favicon'
@@ -13,6 +12,7 @@ import FreeCreditsModal from 'src/components/FreeCreditsModal'
 import IdleStatusMonitor from 'src/components/IdleStatusMonitor'
 import ImportStatus from 'src/components/ImportStatus'
 import { NpsSurvey } from 'src/components/NpsSurvey'
+import { AuthenticatedCookieSetter } from 'src/components/runtime-common'
 import ServiceAlerts from 'src/components/ServiceAlerts'
 import SupportRequest from 'src/components/SupportRequest'
 import { TrialBanner } from 'src/components/TrialBanner'
